refactor: add enum
