+++ blr-map/src/components/Container.tsx
@@ -1,13 +1,13 @@
 import Header from "./Header";
-import Map from './Map';
 import Footer from "./Footer";
 import './styles.scss';
+import MapComponent from "./MapComponent";
 
 const Container = () => {
   return (
     <div className="container">
       <Header />
-      <Map />
+      <MapComponent />
       <Footer />
     </div>
   );
